requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
openai==1.51.0
python-dotenv==1.0.1
pyyaml==6.0.2
gradio==4.44.0
pandas
